% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lines}
\alias{add_lines}
\title{takes a vector of strings, adds another vector
either before or after pattern provided.}
\usage{
add_lines(target, pattern, what, before = TRUE)
}
\arguments{
\item{target}{the vector to add what to}

\item{pattern}{where to add the what to}

\item{what}{vector to add to target}

\item{before}{or after the pattern}
}
\description{
takes a vector of strings, adds another vector
either before or after pattern provided.
}
