% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_to_proj_dir}
\alias{change_to_proj_dir}
\title{Useful function to find temp project name from
temporary file in /tmp/HasHfolDER/.geoplumber.dat
If that is not available (user could already be in a
geoplumber directory) then just returns current wd.}
\usage{
change_to_proj_dir()
}
\description{
Useful function to find temp project name from
temporary file in /tmp/HasHfolDER/.geoplumber.dat
If that is not available (user could already be in a
geoplumber directory) then just returns current wd.
}
