% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_add_geojson.R
\name{gp_add_geojson}
\alias{gp_add_geojson}
\title{For now, views a geojson served from an endpoint on leafletjs.}
\usage{
gp_add_geojson(endpoint = "/api/data")
}
\arguments{
\item{endpoint}{where to fetch the geojson from}
}
\description{
Assumes there is a valid geoplumber app at wd
TODO: work with valid paths too
TODO: either here, or separate function, add my own component.
\enumerate{
\item Add a component with ability to fetch geojson from an endpoint to /components
\item Import it into the App.js (TODO: into any other component)
Errors: stop execution if either React 'src' folder or for now "Welcome.js" component
does not exist. This is because for now we are adding the ready to use GeoJSONComponent component
into src/Welcome.js
}
}
\examples{
\dontrun{
gp_add_geojson()
# Error in gp_add_geojson() : Is current dir a geoplumber app?
# Try geoplumber::gp_create()
}

}
