% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_slider.R
\name{gp_add_slider}
\alias{gp_add_slider}
\title{Adds a basic slider with a callback function to parent.}
\usage{
gp_add_slider(min = 1L, max = 10L, step = 1L,
  js_on_change_function = "onChange={(sliderInput) => this.setState({sliderInput})}",
  to_vector = "NA")
}
\arguments{
\item{min}{min to pass to the slider}

\item{max}{max to pass to the slider}

\item{step}{step changes for min & max}

\item{js_on_change_function}{the function to run on React parent (Welcome). By default,
\code{onChange={(sliderInput) => this.setState({sliderInput})}} sets the state of parent
with value returned from the html input's onChange function.}

\item{to_vector}{instead of reading default Welcome.js}
}
\description{
At this moment, there is no strategy to manage parent <> child
interaction between what is main component (Welcome) and
children. React does have strong tools for this but geoplumber
is still young.
}
\examples{
\dontrun{
gp_add_slider()
}

}
