% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_build.R
\name{gp_build}
\alias{gp_build}
\title{Build npm packages}
\usage{
gp_build(clean = FALSE)
}
\arguments{
\item{clean}{clean build TODO}
}
\description{
This function must be called to build
React build for online front-end.
}
\details{
Runs npm run build.
}
