% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gp_change_file}
\alias{gp_change_file}
\title{Change a source file in place}
\usage{
gp_change_file(
  path = system.file("js/src/App.js", package = "geoplumber"),
  what = " * geoplumber R package code.",
  pattern = " * geoplumber R package code.",
  before = TRUE,
  replace = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path}{path of file to change, used in readLines()}

\item{what}{vector to add to path}

\item{pattern}{where to add the what to, 1st is used. Unique is best.}

\item{before}{s after the pattern}

\item{replace}{or replace pattern}

\item{verbose}{cat the change out}
}
\description{
Utility function to make changes to a source file
}
\examples{
{
 gp_change_file(replace = TRUE, verbose = TRUE) # replacing the comment itself.
}
}
