% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_from_clip.R
\name{gp_check_clip_endpoint}
\alias{gp_check_clip_endpoint}
\title{Basic sanity check of the plumber endpiont}
\usage{
gp_check_clip_endpoint(evaluate = FALSE)
}
\arguments{
\item{evaluate}{should clipboard function be evaulated? Default is (\code{FALSE})}
}
\value{
number of warnings
}
\description{
Use this function to check that:
}
\details{
\enumerate{
\item There is an endpoint "/api/test" etc.
\item There is a "tag" such as @get/@post
\item Defines a function with/without params
\item Serves a content-type https://www.w3.org/TR/html4/types.html#h-6.7.
No specific checks on the return for now just !is.null()
\item TODO: content-type matches
}

using clipr we read from the clipboard
}
\examples{
\dontrun{
gp_check_clip_endpoint()
}
}
