% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{gp_clean}
\alias{gp_clean}
\title{Function to clean the front end build for some reason.}
\usage{
gp_clean(background = FALSE, rebuild = FALSE, reinstall = TRUE)
}
\arguments{
\item{background}{run command in the background,
defaults to \code{FALSE} just to keep user aware.}

\item{rebuild}{whether to rebuild the front, defaults to \code{FALSE}}

\item{reinstall}{should it just clean node_modules & reinstall them,
defaults to \code{TRUE}}
}
\description{
Three options: clean, cleand and reinstall and rebuild
which cleans, reinstall and rebuilds.
}
\examples{
\dontrun{
gp_clean()
}
}
