% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_endpoint_from_clip.R
\name{gp_endpoint_from_clip}
\alias{gp_endpoint_from_clip}
\title{Adds an endpoint function to the plumber.R from clipboard.}
\usage{
gp_endpoint_from_clip(evaluate = FALSE)
}
\arguments{
\item{evaluate}{extra check on clipboard content}
}
\value{
None
}
\description{
To use this function, write the endpoint somewhere and then copy it into clipboard.
Then call this function.
This function uses \code{clipr} to write it to the 'plumber.R' file.
For now only file will be plumber.R
}
\details{
TODO:
add silent param, write to other .R files.
}
\examples{
\dontrun{
# Following is a valid endpoint to serve geoplumber::traffic dataset:
# = begin ===>
# Serve geoplumber::traffic from /api/data

# @get /api/data
# get_traffic <- function(res) {
#   geojson <- geojsonio::geojson_json(geoplumber::traffic)
#   res$body <- geojson
#   res
# }
# <==== end =
# holindg current clipboard
old_clip <- clipr::read_clip()
# adding above to clipboard
clipr::write_clip(c(
"#' Serve geoplumber::traffic from /api/data",
"@get /api/data",
"get_traffic <- function(res) {",
"geojson <- geojsonio::geojson_json(geoplumber::traffic)",
 "res$body <- geojson",
 "res",
"}"
))
# clipr::read_clip()
gp_endpoint_from_clip()
clipr::write_clip(old_clip)
}

}
