% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_erase}
\alias{gp_erase}
\title{Remove a plumber project and clean associated directories}
\usage{
gp_erase(dir_name = NULL)
}
\arguments{
\item{dir_name}{Name of project directory (if NULL, previously-built
directory will be erased)}
}
\description{
Remove a plumber project and clean associated directories
}
