% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{gp_explore}
\alias{gp_explore}
\title{Explore an sf R object using Turing eAtlas.}
\usage{
gp_explore(sf = geoplumber::traffic, build = FALSE, run = TRUE)
}
\arguments{
\item{sf}{a valid sf object that can be converted to geojson}

\item{build}{if \code{TRUE} build the front-end.}

\item{run}{if \code{TRUE} run geoplumber}
}
\description{
Explore an sf R object using Turing eAtlas.
}
\examples{
\dontrun{
gp_explore()
}
}
