% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{gp_geojson}
\alias{gp_geojson}
\title{Explore a geojson object from a remote URL on a map.}
\usage{
gp_geojson(geojson_url, colour_pal)
}
\arguments{
\item{geojson_url}{URL or path to read geojson from}

\item{colour_pal}{the value to use when colouring each feature}
}
\description{
Explore a geojson object from a remote URL on a map.
}
\examples{
\dontrun{
gp_geojson(paste0("http://opendata.canterburymaps.govt.nz/datasets/",
"fb00b553120b4f2fac49aa76bc8d82aa_26.geojson"))
}
}
