% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_node_instructions.R
\name{gp_install_node_instructions}
\alias{gp_install_node_instructions}
\title{TODO: install node for different systems}
\usage{
gp_install_node_instructions()
}
\description{
Currently shows nodejs for deb 64bit distros only
}
\examples{
\dontrun{
gp_install_node_instructions()
}
}
