% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_install_npm_package.R
\name{gp_install_npm_package}
\alias{gp_install_npm_package}
\title{Install an npm package locally}
\usage{
gp_install_npm_package(pkg)
}
\arguments{
\item{pkg}{of npm package to install}
}
\description{
TODO: in future it could do global installation.
}
