% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-react.R
\name{gp_install_react}
\alias{gp_install_react}
\title{Install react npm package}
\usage{
gp_install_react(local = TRUE)
}
\arguments{
\item{local}{should the node modules be installed locally (the default, \code{TRUE})
or globally (\code{FALSE})}
}
\description{
Installs the stack of npm packages needed for building the react app.
Warning: may take several minutes depending on system and internet connection.
}
\examples{
\dontrun{
gp_install_react()
}
}
