% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_is_wd_geoplumber.R
\name{gp_is_wd_geoplumber}
\alias{gp_is_wd_geoplumber}
\title{Helper function to determin if working dir is a geoplumber app.}
\usage{
gp_is_wd_geoplumber(path = ".")
}
\arguments{
\item{path}{check particular path}
}
\value{
true or false
}
\description{
Conditions for a geoplumber app (at least initiated with)
\enumerate{
\item An 'R' directory with R/plumber.R file
\item A 'src' directory
\item A 'package.json' file at root.
}
}
\details{
TODO: what else?
}
\examples{
{
gp_is_wd_geoplumber()
}

}
