% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_npm_exists}
\alias{gp_npm_exists}
\title{Essential checks for certain functions of geoplumber.}
\usage{
gp_npm_exists()
}
\value{
TRUE/FALSE
}
\description{
gp_build, gp_create and others rely on npm/node being present
on the system and might be used in future so refactoring a helper function
is good.
}
\examples{
{
gp_npm_exists()
}
}
