% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb.R
\name{gp_plumb}
\alias{gp_plumb}
\title{Launch a webserver via plumber to serve data}
\usage{
gp_plumb(run = TRUE, port = 8000, file = "R/plumber.R",
  front = FALSE, host = "http://localhost:")
}
\arguments{
\item{run}{should plumber run the server or return it as an object}

\item{port}{to serve from}

\item{file}{location of plumber.R file used by plumber}

\item{front}{should geoplumber start the front dev server? Defaults}

\item{host}{host to pass to plumber
to \code{FALSE}.}
}
\description{
Launch a webserver via plumber to serve data
}
\examples{
\dontrun{
gp_install_react() # install react
gp_build() # build frond-end
gp_plumb()
}
}
\seealso{
\code{\link[=gp_plumb_front]{gp_plumb_front()}}
}
