% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb_front.R
\name{gp_plumb_front}
\alias{gp_plumb_front}
\title{Wrapper function to view front end using dev server.}
\usage{
gp_plumb_front(background = TRUE)
}
\arguments{
\item{background}{run the command in the background, default \code{TRUE}}
}
\description{
Currently geoplumber only works with React,
create-react-app comes with a dev server to view the front end..
This function must be called from a geoplumber app.
}
\examples{
\dontrun{
gp_plumb_front()
}
}
