% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gp_remove_lines}
\alias{gp_remove_lines}
\title{Remove lines from a source file in place}
\usage{
gp_remove_lines(path, pattern = " * geoplumber R package code.",
  lines_count = 1L)
}
\arguments{
\item{path}{path of file to change, used in readLines()}

\item{pattern}{remove what, 1st is used. Unique is best.}

\item{lines_count}{1 by default provide a number}
}
\description{
Utility function to remove lines from a source file
}
\examples{
\dontrun{
 gp_remove_lines()
}
}
