% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gp_rstudio}
\alias{gp_rstudio}
\title{Wrapper function to copy template.Rproj file into working directory.}
\usage{
gp_rstudio(project_name)
}
\arguments{
\item{project_name}{the project name to use for the .Rproj file.}
}
\description{
Wrapper function to copy template.Rproj file into working directory.
}
\examples{
\dontrun{
 gp_rstudio()
}
}
