% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_port_engated}
\alias{is_port_engated}
\title{Internal function to determine if port is engaed.}
\usage{
is_port_engated(port = 3000)
}
\arguments{
\item{port}{to check.}
}
\description{
Internal function to determine if port is engaed.
}
