% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traffic_network}
\alias{traffic_network}
\title{Small sample of road network in an sf object.}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 361 rows and 1 columns.}
\usage{
traffic_network
}
\description{
The data has been processed too much for a reproducible snippet to
provide here.
}
\details{
It is obtained from the same \href{https://data.cdrc.ac.uk/dataset/southwark-traffic-counts}{data.cdrc.ac.uk/dataset/southwark-traffic-counts}
dataset of \code{traffic} above, also using osmdata package.
}
\examples{
{
object.size(traffic_network)
}
}
\keyword{datasets}
