% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traffic_volumes}
\alias{traffic_volumes}
\title{Traffic data from CDRC data sets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1522 rows and 20 columns.
}
\usage{
traffic_volumes
}
\description{
See \href{https://data.cdrc.ac.uk/dataset/southwark-traffic-counts}{data.cdrc.ac.uk/dataset/southwark-traffic-counts}
This dataset represents travel levels at the point locations stored in
\code{traffic}
}
\examples{
\dontrun{
f = "/tmp/2010-2016-allaveragedailytrafficflowcdrcmapreducedfile.xlsx"
traffic_volumes = readxl::read_excel(f)
summary(traffic_volumes$ID \%in\% traffic$cp)
pryr::object_size(traffic_volumes)
usethis::use_data(traffic_volumes)
library(dplyr)
traffic_agg = traffic_volumes \%>\%
  group_by(cp = ID) \%>\%
  summarise(
    total = sum(`TOTAL FLOW`),
    cycle_or_motorcycle = sum(`ARX PEDAL-MOTORCYCLE`),
    av_speed = mean(AVERAGE_SPEED),
    year = mean(YEAR)
    )
if(ncol(traffic) < 8) {
  traffic = left_join(traffic, traffic_agg)
  usethis::use_data(traffic, overwrite = TRUE)
}
mapview::mapview(traffic, zcol = "total")
sf:::plot.sf(traffic)
}
summary(traffic_volumes)
plot(traffic$total, traffic$cycle_or_motorcycle)
}
\keyword{datasets}
