% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uni_point}
\alias{uni_point}
\alias{uni_poly}
\title{sf object representing the university of Leeds}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 5 columns.
}
\usage{
uni_point
}
\description{
From OpenStreetMap
}
\examples{
# library(osmdata)
# osm_list <- opq("leeds") \%>\%
#   add_osm_feature("name", "University of Leeds") \%>\%
#   osmdata_sf()
# uni_poly = osm_list$osm_polygons
# mapview::mapview(uni_poly)
# uni_point <- sf::st_centroid(osm_list$osm_polygons)
}
\keyword{datasets}
