% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_import_component}
\alias{add_import_component}
\title{takes a vector of strings, adds a Babel style import statement}
\usage{
add_import_component(target, component.name, component.path)
}
\arguments{
\item{target}{vector to add import in}

\item{component.name}{Raect name of component}

\item{component.path}{path to "import" from}
}
\description{
takes a vector of strings, adds a Babel style import statement
}
