% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_slider.R
\name{gp_add_slider}
\alias{gp_add_slider}
\title{Adds a basic slider with a callback function to parent.}
\usage{
gp_add_slider()
}
\description{
Adds a basic slider with a callback function to parent.
}
\examples{
\dontrun{
gp_add_slider()
}

}
