% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_cra_init}
\alias{gp_cra_init}
\title{Simulate CRA without create-react-app}
\usage{
gp_cra_init(path = getwd())
}
\arguments{
\item{path}{character: existing path of the target gp app.}
}
\description{
Because \code{[gp_create()]} uses the underlying CRA npm package, it is slow.
This function assembles the required npm package files to then build from.
}
\examples{
gp_cra_init(tempdir()) # it would simulate an app in tempdir()
}
