% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_create}
\alias{gp_create}
\title{Simulate CRA without create-react-app}
\usage{
gp_create(path = getwd())
}
\arguments{
\item{path}{character: existing path of the target gp app.}
}
\description{
This function assembles the required npm package files to then build from.
}
\examples{
gp_create(tempdir()) # it would simulate an app in tempdir()
}
