% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_create}
\alias{gp_create}
\title{Initialize a new create-react-app and add add required packages.}
\usage{
gp_create(path = "geoplumber")
}
\arguments{
\item{path}{path of the new app, will be passed to \code{npx create-react-app }}
}
\description{
Warning: may take several minutes depending on system and internet connection.
}
\examples{
\dontrun{
gp_create()
}
}
