% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{gp_geojson}
\alias{gp_geojson}
\title{Explore a geojson object from a remote URL a leaflet map.}
\usage{
gp_geojson(url)
}
\arguments{
\item{url}{path to pull the geojson from}
}
\description{
Explore a geojson object from a remote URL a leaflet map.
}
\examples{
\dontrun{
gp_geojson("http://opendata.canterburymaps.govt.nz/datasets/fb00b553120b4f2fac49aa76bc8d82aa_26.geojson")
}
}
