% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb.R
\name{gp_plumb}
\alias{gp_plumb}
\title{Launch a webserver via plumber to serve data}
\usage{
gp_plumb(
  run = TRUE,
  port = 8000,
  file = "R/plumber.R",
  front = FALSE,
  host = "127.0.0.1",
  background = TRUE
)
}
\arguments{
\item{run}{should plumber run the server or return it as an object}

\item{port}{to serve from}

\item{file}{location of plumber.R file used by plumber}

\item{front}{should geoplumber start the front dev server? Defaults}

\item{host}{host to pass to plumber default \verb{http://127.0.0.1}
to \code{FALSE}.}

\item{background}{run the R process in the background using callr,
defaults to \code{TRUE}.}
}
\value{
an instance of plumber::plumb if \code{run} is set to \code{FALSE},
a process id from \code{callr::r_bg} to easily destroy if parameter
\code{background} is default \code{TRUE}.
}
\description{
Launch a webserver via plumber to serve data
}
\examples{
{
d <- file.path(tempdir(), "gp")
gp_create(d)
old <- setwd(d)
ps <- gp_plumb()
ps
Sys.sleep(1) # just in case
ps
require(RCurl)
webpage <- getURL("http://localhost:8000")
webpage <- readLines(tc <- textConnection(webpage)); close(tc)
tail(webpage)
ps$kill()
setwd(old)
unlink(d, recursive = TRUE)
}
}
\seealso{
\code{\link[=gp_plumb_front]{gp_plumb_front()}}
}
