% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_plumb.R
\name{gp_plumb}
\alias{gp_plumb}
\title{Launch a webserver via plumber to serve data}
\usage{
gp_plumb(port = 8000, file = "R/plumber.R")
}
\arguments{
\item{port}{to serve from}

\item{file}{location of plumber.R file used by plumber}
}
\description{
Launch a webserver via plumber to serve data
}
\examples{
\dontrun{
gp_install_react() # install react
gp_build() # build frond-end
gp_plumb()
}
}
