% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_comp.R
\name{gp_react_comp}
\alias{gp_react_comp}
\title{Creates an R6 R represetnation of a ReactJS component}
\usage{
gp_react_comp(name = "R", props = list(r = "R6"))
}
\arguments{
\item{name}{React component name}

\item{props}{a named R list of React props for the component.}
}
\description{
WIP
}
\examples{
{
library(R6)
r <- gp_react_comp("helloworld")
r
# React Component:
# import React, { Component } from 'react';
# export default class helloworld extends Component {
#   render() {
#     return(
#       <p>
#         I am helloworld.
#       </p>
#     )
#   }
# }
}
}
