% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gp_rstudio}
\alias{gp_rstudio}
\title{Wrapper function to copy template.Rproj file into working directory.}
\usage{
gp_rstudio(path = ".")
}
\arguments{
\item{path}{project path to create .Rproj file in, defaults to ".".}
}
\description{
Wrapper function to copy template.Rproj file into working directory.
}
\examples{
\dontrun{
 gp_rstudio()
}
}
