% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{gp_sf}
\alias{gp_sf}
\title{Explore an sf R object on a leaflet map.}
\usage{
gp_sf(
  sf = geoplumber::traffic,
  props_list = list(road = geoplumber::traffic$road)
)
}
\arguments{
\item{sf}{a valid sf object that can be converted to geojson}

\item{props_list}{one named list of menuitmes to explore sf object with.}
}
\description{
The function (for now) takes one parameter to bounce back
to the backend. For now just a dropdown list.
The slider only works with circles.
}
\examples{
\dontrun{
gp_sf()
}
}
